--- cmd/influxd/run/server_default.go
@@ -1,149 +1,0 @@
-// +build !dataonly
-
-package run
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-
-	"github.com/influxdata/influxdb/services/admin"
-	"github.com/influxdata/influxdb/services/collectd"
-	"github.com/influxdata/influxdb/services/continuous_querier"
-	"github.com/influxdata/influxdb/services/graphite"
-	"github.com/influxdata/influxdb/services/httpd"
-	"github.com/influxdata/influxdb/services/meta"
-	"github.com/influxdata/influxdb/services/opentsdb"
-	"github.com/influxdata/influxdb/services/precreator"
-	"github.com/influxdata/influxdb/services/retention"
-	"github.com/influxdata/influxdb/services/udp"
-)
-
-func (s *Server) appendMonitorService() {
-	s.Services = append(s.Services, s.Monitor)
-}
-
-func (s *Server) appendRetentionPolicyService(c retention.Config) {
-	if !c.Enabled {
-		return
-	}
-	srv := retention.NewService(c)
-	srv.MetaClient = s.MetaClient
-	srv.TSDBStore = s.TSDBStore
-	s.Services = append(s.Services, srv)
-}
-
-func (s *Server) appendAdminService(c admin.Config) {
-	if !c.Enabled {
-		return
-	}
-	c.Version = s.buildInfo.Version
-	srv := admin.NewService(c)
-	s.Services = append(s.Services, srv)
-}
-
-func (s *Server) appendHTTPDService(c httpd.Config) {
-	if !c.Enabled {
-		return
-	}
-	srv := httpd.NewService(c)
-	srv.Handler.MetaClient = s.MetaClient
-	srv.Handler.QueryAuthorizer = meta.NewQueryAuthorizer(s.MetaClient)
-	srv.Handler.WriteAuthorizer = meta.NewWriteAuthorizer(s.MetaClient)
-	srv.Handler.QueryExecutor = s.QueryExecutor
-	srv.Handler.PointsWriter = s.PointsWriter
-	srv.Handler.Version = s.buildInfo.Version
-
-	// If a ContinuousQuerier service has been started, attach it.
-	for _, srvc := range s.Services {
-		if cqsrvc, ok := srvc.(continuous_querier.ContinuousQuerier); ok {
-			srv.Handler.ContinuousQuerier = cqsrvc
-		}
-	}
-
-	s.Services = append(s.Services, srv)
-}
-
-func (s *Server) appendCollectdService(c collectd.Config) {
-	if !c.Enabled {
-		return
-	}
-	srv := collectd.NewService(c)
-	srv.MetaClient = s.MetaClient
-	srv.PointsWriter = s.PointsWriter
-	s.Services = append(s.Services, srv)
-}
-
-func (s *Server) appendOpenTSDBService(c opentsdb.Config) error {
-	if !c.Enabled {
-		return nil
-	}
-	srv, err := opentsdb.NewService(c)
-	if err != nil {
-		return err
-	}
-	srv.PointsWriter = s.PointsWriter
-	srv.MetaClient = s.MetaClient
-	s.Services = append(s.Services, srv)
-	return nil
-}
-
-func (s *Server) appendGraphiteService(c graphite.Config) error {
-	if !c.Enabled {
-		return nil
-	}
-	srv, err := graphite.NewService(c)
-	if err != nil {
-		return err
-	}
-
-	srv.PointsWriter = s.PointsWriter
-	srv.MetaClient = s.MetaClient
-	srv.Monitor = s.Monitor
-	s.Services = append(s.Services, srv)
-	return nil
-}
-
-func (s *Server) appendPrecreatorService(c precreator.Config) error {
-	if !c.Enabled {
-		return nil
-	}
-	srv, err := precreator.NewService(c)
-	if err != nil {
-		return err
-	}
-
-	srv.MetaClient = s.MetaClient
-	s.Services = append(s.Services, srv)
-	return nil
-}
-
-func (s *Server) appendUDPService(c udp.Config) {
-	if !c.Enabled {
-		return
-	}
-	srv := udp.NewService(c)
-	srv.PointsWriter = s.PointsWriter
-	srv.MetaClient = s.MetaClient
-	s.Services = append(s.Services, srv)
-}
-
-func (s *Server) appendContinuousQueryService(c continuous_querier.Config) {
-	if !c.Enabled {
-		return
-	}
-	srv := continuous_querier.NewService(c)
-	srv.MetaClient = s.MetaClient
-	srv.QueryExecutor = s.QueryExecutor
-	s.Services = append(s.Services, srv)
-}
-
-func raftDBExists(dir string) error {
-	// Check to see if there is a raft db, if so, error out with a message
-	// to downgrade, export, and then import the meta data
-	raftFile := filepath.Join(dir, "raft.db")
-	if _, err := os.Stat(raftFile); err == nil {
-		return fmt.Errorf("detected %s. To proceed, you'll need to either 1) downgrade to v0.11.x, export your metadata, upgrade to the current version again, and then import the metadata or 2) delete the file, which will effectively reset your database. For more assistance with the upgrade, see: https://docs.influxdata.com/influxdb/v0.12/administration/upgrading/", raftFile)
-	}
-	return nil
-}
